move menu from user_choice to main
